+++ src/main/java/bigworkers/ingresoegreso/ferreteria/service/IDocumentTypeService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface IDocumentTypeService {
 
 
-    public DocumentType findById(long id);
+    public DocumentType findById(int id);
 
 
     public List<DocumentType> findAll();
@@ -15,9 +15,9 @@ public interface IDocumentTypeService {
     public DocumentType createDocumentType(DocumentType tipoDocumento);
 
 
-    public DocumentType updateDocumentType(long id, DocumentType tipoDocumento);
+    public DocumentType updateDocumentType(int id, DocumentType tipoDocumento);
 
 
-    public void deleteDocumentType(long id);
+    public void deleteDocumentType(int id);
 
 }
