+++ src/main/java/bigworkers/ingresoegreso/ferreteria/service/IProfileService.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public interface IProfileService {
 
-    public Profile findById(long id);
+    public Profile findById(int id);
 
     public List<Profile> findAll();
 
     public Profile createProfile(Profile perfil);
 
-    public Profile updateProfile(long id,Profile perfil);
+    public Profile updateProfile(int id,Profile perfil);
 
-    public void deleteProfile(long id);
+    public void deleteProfile(int id);
 }
